+++ CHANGELOG.md
@@ -13,6 +13,7 @@ If you like this project and find it useful, please consider giving it a star on
 ### Added
 
 - [jsdoc]: Improved explanation on jsdoc cluster helpers.
+- [cover]: Added createDefaultTiltWindowCoveringClusterServer().
 
 ### Changed